Update branch renaming guidance
This is to add a warning and a new step. The warning is to check branch protection rules, as they don't get updated, so you could inadvertently allow pushing to your new default branch. The new step is to preserve any open PR's as per https://stackoverflow.com/questions/20007578/renaming-a-branch-while-on-pull-request.